pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool.poetry]
authors = ["Tomas Coufal <tcoufal@redhat.com>"]
classifiers = [
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Testing",
]
description = "Account Management Tool for Testing"
keywords = ["testing-tool", "account-management", "candlepin"]
license = "MIT"
maintainers = ["Tomas Coufal <tcoufal@redhat.com>"]
name = "ethel"
readme = "README.md"
repository = "https://github.com/tumido/ethel"
version = "3.0.0.rc2"

[tool.poetry.dependencies]
python = "^3.7"
requests = "*"
pyyaml = "*"
Jinja2 = "*"

[tool.poetry.dev-dependencies]
ipython = "*"
mypy = "*"
pylint = {git = "https://github.com/PyCQA/pylint.git"}  # Use GIT version until 2.5.0 with pyproject.toml support is released
pytest = "^5.3.5"
taskipy = "^1.1.3"
safety = "^1.8.5"
pytest-vcr = "^1.0.2"
hypothesis = "^5.5.2"
pytest-mock = "^2.0.0"
black = "^19.10b0"
pytest-cov = "^2.8.1"

[tool.taskipy.tasks]
audit = "safety check"
lint = "mypy .; pylint ethel tests"
test = "pytest --cov=ethel tests"
test-ci = "pytest --cov=ethel --vcr-record=none tests"

[tool.black]
line-length = 90

[tool.pylint.basic]
good-names = "e"

[tool.pylint.format]
max-line-length = "90"

[tool.pylint.design]
max-args = "10"

[tool.pylint.messages_control]
disable = """
    bad-continuation,duplicate-code,
    missing-module-docstring,missing-class-docstring,
    too-many-instance-attributes,too-few-public-methods
"""
